chore: Break out std functions to dedicated modules.
